Bug 1741018 - Use uint32_t for bfcache flags and MOZ_LOG for ACTIVE_LOCK r=peterv
Differential Revision: https://phabricator.services.mozilla.com/D131523

UltraBlame original commit: e476f8494a849e440b6e49cd443d0370bc3987b8

+++ docshell/base/nsDocShell.cpp
@@ -6804,13 +6804,13 @@ bool nsDocShell::CanSavePresentation(uint32_t aLoadType,
 }
 
 
-void nsDocShell::ReportBFCacheComboTelemetry(uint16_t aCombo) {
+void nsDocShell::ReportBFCacheComboTelemetry(uint32_t aCombo) {
   
   
   
   
   
-  enum BFCacheStatusCombo : uint16_t {
+  enum BFCacheStatusCombo : uint32_t {
     BFCACHE_SUCCESS,
     NOT_ONLY_TOPLEVEL = mozilla::dom::BFCacheStatus::NOT_ONLY_TOPLEVEL_IN_BCG,
